Refactor NPC Hierarcy
Changelog:
-Made AbstractNPC extend PathfinderMob and Implenent NeutralMob.
- Moved duplicate code in AbstractPirateNPC and AbstractNavyNPC to AbstractNPC, and made AbstractPirate/NavyNPC extend Abstract NPC.

-Also removed some unneeded imports.

-Still have more to refactor
