+++ src/main/java/net/minemod/onepiecemod/entity/npcs/AbstractNPC.java
@@ -1,12 +1,32 @@
 package net.minemod.onepiecemod.entity.npcs;
 
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.util.TimeUtil;
+import net.minecraft.util.valueproviders.UniformInt;
 import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.NeutralMob;
 import net.minecraft.world.entity.PathfinderMob;
 import net.minecraft.world.entity.ai.goal.*;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
+import org.jspecify.annotations.Nullable;
 
-public class AbstractNPC extends PathfinderMob {
+import java.util.UUID;
+
+public abstract class AbstractNPC extends PathfinderMob implements NeutralMob {
+
+    // TODO: are these needed in AbstractNPC? or will they need to be redeclared?
+
+
+    /** The remaining time (in ticks) that this NPC will stay angry at its target. */
+    private int remainingPersistentAngerTime;
+
+    /** The UUID of the player or entity that this NPC is currently angry at. */
+    private UUID persistentAngerTarget;
+
+    /** Defines the range of time (20 to 39 seconds) that anger will persist when triggered. */
+    private static final UniformInt PERSISTENT_ANGER_TIME = TimeUtil.rangeOfSeconds(20,39);
+
 
     public AbstractNPC(EntityType<? extends PathfinderMob> type, Level pLevel) {
         super(type, pLevel);
@@ -19,15 +39,14 @@ public class AbstractNPC extends PathfinderMob {
      */
     @Override
     protected void registerGoals() {
-        // (if player has devil fruit, sink has priority 0)
+        // Combat behavior
+        this.goalSelector.addGoal(0, new MeleeAttackGoal(this, 1.2D, true)); // Attacks target when in range
+
+        // Movement and idle behavior
         this.goalSelector.addGoal(1, new FloatGoal(this)); // Allows swimming
-
-        this.goalSelector.addGoal(2, new PanicGoal(this, 2.0)); // When hit
-        this.goalSelector.addGoal(2, new RandomStrollGoal(this, 1.0D)); // Wandering
-
-        this.goalSelector.addGoal(3, new WaterAvoidingRandomStrollGoal(this, 1.0F));
-        this.goalSelector.addGoal(4, new LookAtPlayerGoal(this, Player.class, 8.0F)); // Looks at players
-        this.goalSelector.addGoal(5, new RandomLookAroundGoal(this)); // Idle head movement
+        this.goalSelector.addGoal(2, new WaterAvoidingRandomStrollGoal(this, 1.0D)); // Wanders around
+        this.goalSelector.addGoal(3, new LookAtPlayerGoal(this, Player.class, 8.0F)); // Looks at nearby players
+        this.goalSelector.addGoal(4, new RandomLookAroundGoal(this)); // Idle head movement
     }
 
     /**
@@ -39,6 +58,46 @@ public class AbstractNPC extends PathfinderMob {
     @Override
     public boolean isPushable() {
         return true;
+    }
+
+    @Override
+    public int getRemainingPersistentAngerTime() {
+        return this.remainingPersistentAngerTime;
+    }
+
+    @Override
+    public void setRemainingPersistentAngerTime(int pRemainingPersistentAngerTime) {
+        this.remainingPersistentAngerTime = pRemainingPersistentAngerTime;
+    }
+
+    @Override
+    public @Nullable UUID getPersistentAngerTarget() {
+        return this.persistentAngerTarget;
+    }
+
+    @Override
+    public void setPersistentAngerTarget(@Nullable UUID pPersistentAngerTarget) {
+        this.persistentAngerTarget = pPersistentAngerTarget;
+    }
+
+    /**
+     * Starts the persistent anger timer by picking a random duration within the defined range.
+     */
+    @Override
+    public void startPersistentAngerTimer() {
+        this.setRemainingPersistentAngerTime(PERSISTENT_ANGER_TIME.sample(this.random));
+    }
+
+    /**
+     * Handles server-side AI steps. Updates the persistent anger timers for the NeutralMob mechanics.
+     */
+    @Override
+    protected void customServerAiStep(ServerLevel pLevel) {
+        // Access pLevel directly
+        this.updatePersistentAnger(pLevel, true);
+
+        // Pass parameter to super method
+        super.customServerAiStep(pLevel);
     }
 
 }
